policy

+++ app/Http/Controllers/Screencast/PlaylistController.php
@@ -46,6 +46,9 @@ class PlaylistController extends Controller
 
    public function edit(Playlist $playlist)
    {
+      // Policy
+      $this->authorize('update', $playlist);
+
       return view('playlists.edit', [
          'playlist' => $playlist,
          'tags'     => Tag::get(),
@@ -54,6 +57,9 @@ class PlaylistController extends Controller
 
    public function update(PlaylistRequest $request, Playlist $playlist)
    {
+      // Policy
+      $this->authorize('update', $playlist);
+
       if ($request->thumbnail) {
          Storage::delete($playlist->thumbnail);
          $thumbnail = $request->file('thumbnail')->store('images/playlist');
@@ -75,6 +81,9 @@ class PlaylistController extends Controller
 
    public function destroy(Playlist $playlist)
    {
+      // Policy
+      $this->authorize('delete', $playlist);
+
       Storage::delete($playlist->thumbnail);
       $playlist->tags()->detach();
       $playlist->delete();
